+++ src/Hellion.World/Managers/FormulasManager.cs
@@ -1,0 +1,26 @@
+﻿using Hellion.World.Systems;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Hellion.World.Managers
+{
+    public static class FormulasManager
+    {
+        public static int GetHpRecovery(Mover mover)
+        {
+            return 0;
+        }
+
+        public static int GetMpRecovery(Mover mover)
+        {
+            return 0;
+        }
+
+        public static int GetFpRecovery(Mover mover)
+        {
+            return 0;
+        }
+    }
+}
